+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `@lumx/react`:
     -   `UserBlock`: improve name and fields text layout in constrained space
+    -   `Slideshow`: add mode prop to switch to a native scroll snap slideshow
 
 ## [4.1.0][] - 2026-01-22